把所有float改为了double，修复了edit下不足一页数据出现的光标异常，复制到缓冲数据的异常，也修复了del下结尾提示不对,增加了array_list找最大元素的方法

+++ proj/data.h
@@ -6,11 +6,10 @@
 // 监测数据结构体
 struct WaterQuality {
    int id; // 记录唯一id
-   // TODO: 改double
-   float tmp; // 水温 度
-   float doxygen; // 溶解氧 mg/L
-   float ph; // PH值 pH
-   float ammonia; // 氨氮含量 mg/L
+   double tmp; // 水温 度
+   double doxygen; // 溶解氧 mg/L
+   double ph; // PH值 pH
+   double ammonia; // 氨氮含量 mg/L
    char time[20]; // 时间记录字符串，格式 yyyy-MM-dd HH:mm:ss
 };
 // 用于指示上面的一部分字段的
